classifier-agent/requirements.txt
# Google ADK for Gemini AI
google-genai>=1.0.0

# FastAPI web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# PDF processing
PyMuPDF>=1.24.0
Pillow>=10.0.0

# Data validation
pydantic>=2.9.0
pydantic-settings>=2.5.0

# HTTP client
httpx>=0.27.0
